fix(OrderRepository): type of id

+++ src/main/java/com/codewithmosh/store/repositories/OrderRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface OrderRepository extends JpaRepository<Order, Byte> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
     @EntityGraph(attributePaths = "items.product")
     @Query("SELECT o FROM Order o WHERE o.customer = :customer")
     List<Order> getOrdersByCustomer(@Param("customer") User customer);
